A naive attempt at implementing dynamic extents/windings.
This approach does not work. This is likely impossible to do with the
way Lillisp is implemented currently on .NET. I think it would take
making Lillisp either a bytecode interpreter (like BiwaScheme) or
a compiler (like Chicken Scheme) where either GOTO or compiled state
machines could be used. But pushing up this branch for future posterity
just in case.